+++ src/main/java/racingcar/domain/NumericGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 public interface NumericGenerator {
     int generate();
