+++ src/main/java/racingcar/domain/numericgenerator/RandomNumericGenerator.java
@@ -1,8 +1,8 @@
-package racingcar.model.numericgenerator;
+package racingcar.domain.numericgenerator;
 
 import java.util.Random;
 
-import racingcar.model.NumericGenerator;
+import racingcar.domain.NumericGenerator;
 
 public class RandomNumericGenerator implements NumericGenerator {
     private final Random random = new Random();
